+++ packages/shared-ui/src/index.ts
@@ -138,7 +138,6 @@ export {
   ShellDropdown,
   type ShellDropdownOption,
 } from './components/shell-dropdown';
-export { ShellNavCard } from './components/shell-nav-card';
 export { ShellNavRow } from './components/shell-nav-row';
 export { ShellFooter } from './components/shell-footer';
 export {
